created voters routes
