+++ db/schema.sql
@@ -1,5 +1,6 @@
 DROP TABLE IF EXISTS parties;
 DROP TABLE IF EXISTS candidates;
+DROP TABLE IF EXISTS voters;
 
 CREATE TABLE parties (
   id INTEGER PRIMARY KEY,
@@ -23,4 +24,12 @@ CREATE TABLE candidates (
   industry_connected BOOLEAN NOT NULL,
   party_id INTEGER CHECK (party_id > 0),
   FOREIGN KEY (party_id) REFERENCES parties(id) ON DELETE SET NULL
+);
+
+CREATE TABLE voters (
+  id INTEGER PRIMARY KEY,
+  first_name VARCHAR(30) NOT NULL,
+  last_name VARCHAR(30) NOT NULL,
+  email VARCHAR(50) NOT NULL,
+  created_at DATETIME DEFAULT CURRENT_TIMESTAMP
 );
