+++ routes/apiRoutes/index.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const router = express.Router();
+
+router.use(require('./candidateRoutes'));
+router.use(require('./partyRoutes'));
+router.use(require('./voterRoutes'));
+
+module.exports = router;
